+++ app/extensions/sample_extension.py
@@ -11,7 +11,7 @@ from app.managers.entity_manager import EntityManager
 from app.managers.cache_manager import CacheManager
 
 
-async def after_startup(
+async def on_startup(
     entity_manager: EntityManager,
     cache_manager: CacheManager, request: None,
     current_user: None,
@@ -23,6 +23,15 @@ async def after_startup(
     function is invoked after the application starts and does not
     return anything.
     """
+    ...
+
+
+async def on_schedule(
+    entity_manager: EntityManager,
+    cache_manager: CacheManager, request: None,
+    current_user: None,
+    entity: None
+):
     ...
 
 
@@ -158,9 +167,9 @@ async def before_collection_insert(
 ) -> Collection:
     """
     Executes additional logic before a collection entity is created,
-    such as logging, updating the database, managing the cache,
+    such as event logging, updating the database, managing the cache,
     performing file or network operations, or other actions. It takes
-    the entity as input, applies the required pre-processing, and then
+    the entity as input, applies the required pre-processing, and
     returns the possibly modified entity.
     """
     ...
@@ -176,9 +185,9 @@ async def after_collection_insert(
 ) -> Collection:
     """
     Executes additional logic after a collection entity is created,
-    such as logging, updating the database, managing the cache,
+    such as event logging, updating the database, managing the cache,
     performing file or network operations, or other actions. It takes
-    the entity as input, applies the required post-processing, and then
+    the entity as input, applies the required post-processing, and
     returns the possibly modified entity.
     """
     ...
@@ -192,6 +201,13 @@ async def after_collection_select(
     current_user: User,
     collection: Collection
 ) -> Collection:
+    """
+    Executes additional logic after a collection entity is selected,
+    such as event logging, updating the database, managing the cache,
+    performing file or network operations, or other actions. It takes
+    the entity as input, applies the required post-processing, and
+    returns the possibly modified entity.
+    """
     ...
     return collection
 
@@ -203,6 +219,13 @@ async def before_collection_update(
     current_user: User,
     collection: Collection
 ) -> Collection:
+    """
+    Executes additional logic before a collection entity is updated,
+    such as event logging, updating the database, managing the cache,
+    performing file or network operations, or other actions. It takes
+    the entity as input, applies the required pre-processing, and
+    returns the possibly modified entity.
+    """
     ...
     return collection
 
@@ -214,6 +237,13 @@ async def after_collection_update(
     current_user: User,
     collection: Collection
 ) -> Collection:
+    """
+    Executes additional logic after a collection entity is updated,
+    such as event logging, updating the database, managing the cache,
+    performing file or network operations, or other actions. It takes
+    the entity as input, applies the required post-processing, and
+    returns the possibly modified entity.
+    """
     ...
     return collection
 
@@ -225,6 +255,13 @@ async def before_collection_delete(
     current_user: User,
     collection: Collection
 ) -> Collection:
+    """
+    Executes additional logic before a collection entity is deleted,
+    such as event logging, updating the database, managing the cache,
+    performing file or network operations, or other actions. It takes
+    the entity as input, applies the required pre-processing, and
+    returns the possibly modified entity.
+    """
     ...
     return collection
 
@@ -236,6 +273,13 @@ async def after_collection_delete(
     current_user: User,
     collection: Collection
 ) -> Collection:
+    """
+    Executes additional logic after a collection entity is deleted,
+    such as event logging, updating the database, managing the cache,
+    performing file or network operations, or other actions. It takes
+    the entity as input, applies the required post-processing, and
+    returns the possibly modified entity.
+    """
     ...
     return collection
 
@@ -249,14 +293,32 @@ async def after_collections_list(
 ) -> List[Collection]:
     """
     Executes additional logic after a list of collection entities is
-    selected, such as logging, updating the database, managing the
-    cache, performing file or network operations, or other actions.
+    selected, such as event logging, updating the database, managing
+    the cache, performing file or network operations, or other actions.
     It takes the list of entities as input, applies the required
-    post-processing, and then returns the list of possibly modified
+    post-processing, and returns the list of possibly modified
     entities.
     """
     ...
     return collections
+
+
+async def before_document_insert(
+    entity_manager: EntityManager,
+    cache_manager: CacheManager,
+    request: Request,
+    current_user: User,
+    document: Document
+) -> Document:
+    """
+    Executes additional logic before a document entity is created,
+    such as event logging, updating the database, managing the cache,
+    performing file or network operations, or other actions. It takes
+    the entity as input, applies the required pre-processing, and
+    returns the possibly modified entity.
+    """
+    ...
+    return document
 
 
 async def after_document_insert(
@@ -266,6 +328,13 @@ async def after_document_insert(
     current_user: User,
     document: Document
 ) -> Document:
+    """
+    Executes additional logic after a document entity is created,
+    such as event logging, updating the database, managing the cache,
+    performing file or network operations, or other actions. It takes
+    the entity as input, applies the required post-processing, and
+    returns the possibly modified entity.
+    """
     ...
     return document
 
@@ -277,6 +346,31 @@ async def after_document_select(
     current_user: User,
     document: Document
 ) -> Document:
+    """
+    Executes additional logic after a document entity is selected,
+    such as event logging, updating the database, managing the cache,
+    performing file or network operations, or other actions. It takes
+    the entity as input, applies the required post-processing, and
+    returns the possibly modified entity.
+    """
+    ...
+    return document
+
+
+async def before_document_update(
+    entity_manager: EntityManager,
+    cache_manager: CacheManager,
+    request: Request,
+    current_user: User,
+    document: Document
+) -> Document:
+    """
+    Executes additional logic before a document entity is updated,
+    such as event logging, updating the database, managing the cache,
+    performing file or network operations, or other actions. It takes
+    the entity as input, applies the required pre-processing, and
+    returns the possibly modified entity.
+    """
     ...
     return document
 
@@ -288,6 +382,31 @@ async def after_document_update(
     current_user: User,
     document: Document
 ) -> Document:
+    """
+    Executes additional logic after a document entity is updated,
+    such as event logging, updating the database, managing the cache,
+    performing file or network operations, or other actions. It takes
+    the entity as input, applies the required post-processing, and
+    returns the possibly modified entity.
+    """
+    ...
+    return document
+
+
+async def before_document_delete(
+    entity_manager: EntityManager,
+    cache_manager: CacheManager,
+    request: Request,
+    current_user: User,
+    document: Document
+) -> Document:
+    """
+    Executes additional logic before a document entity is deleted,
+    such as event logging, updating the database, managing the cache,
+    performing file or network operations, or other actions. It takes
+    the entity as input, applies the required pre-processing, and
+    returns the possibly modified entity.
+    """
     ...
     return document
 
@@ -299,6 +418,13 @@ async def after_document_delete(
     current_user: User,
     document: Document
 ) -> Document:
+    """
+    Executes additional logic after a document entity is deleted,
+    such as event logging, updating the database, managing the cache,
+    performing file or network operations, or other actions. It takes
+    the entity as input, applies the required post-processing, and
+    returns the possibly modified entity.
+    """
     ...
     return document
 
@@ -310,6 +436,14 @@ async def after_documents_list(
     current_user: User,
     documents: List[Document]
 ) -> List[Document]:
+    """
+    Executes additional logic after a list of document entities is
+    selected, such as event logging, updating the database, managing
+    the cache, performing file or network operations, or other actions.
+    It takes the list of entities as input, applies the required
+    post-processing, and returns the list of possibly modified
+    entities.
+    """
     ...
     return documents